헥사고날 변경에 따른 프론트 변경

+++ src/service/categoryService.js
@@ -8,11 +8,11 @@ const categoryService = {
         return axiosClient.get(`/api/category/root`)
     },
     save: ({category}) => {
-        const {id, name, pId} = category
-        return axiosClient.post(`/api/category`, {id: id, name: name, pId: pId})
+        const {id, name, parentId} = category
+        return axiosClient.post(`/api/category/admin`, {id: id, name: name, parentId: parentId})
     },
     delete: ({id}) => {
-        return axiosClient.delete(`/api/category/${id}`)
+        return axiosClient.delete(`/api/category/admin/${id}`)
     }
 }
